スキルの関数の調整
