+++ components/top/Skils.tsx
@@ -1,93 +1,16 @@
+import { showSlils, skilsOptions } from '@/hooks/top/skils';
+import { useIntersectionObserver } from '@/hooks/use-intersection-observer';
 import { TopProps } from '@/types/types';
 import Image from 'next/image';
-import React, { FC } from 'react';
+import React, { FC, useRef } from 'react';
 
 const Skils: FC<TopProps> = ({ item }) => {
-  if (typeof window === 'object') {
-    const progressItems = document.querySelectorAll('.progress-item');
+  const s1 = useRef<HTMLDivElement>(null);
+  const s2 = useRef<HTMLDivElement>(null);
+  const s3 = useRef<HTMLDivElement>(null);
+  const s4 = useRef<HTMLDivElement>(null);
 
-    const root = document.querySelector('#top');
-
-    const observeAction = (entries: any) => {
-      entries.forEach((entry: any) => {
-        if (
-          !entry.isIntersecting ||
-          entry.target.classList.contains('is-visible')
-        ) {
-          return;
-        }
-
-        const percent = parseFloat(entry.target.dataset.percent);
-
-        if (!Number.isFinite(percent)) {
-          return false;
-        }
-
-        const duration = parseInt(entry.target.dataset.duration) || 1500;
-        const eleProgressBar = entry.target.querySelector('.progress-bar');
-        const eleProgressValue = entry.target.querySelector('.progress-value');
-        const radius = eleProgressBar.getAttribute('r');
-        const circumference = 2 * Math.PI * radius;
-        const strokeDashOffset = Math.round(
-          circumference - (circumference * percent) / 100
-        );
-
-        // Progress start position
-        let startPosition: { [key: string]: string };
-
-        startPosition = {
-          right: '0deg',
-          bottom: '90deg',
-          left: '180deg',
-          default: '-90deg',
-        };
-
-        startPosition[entry.target.dataset.startPosition] || '-90deg';
-
-        // Countup percentage
-        const deciamlPointLength = (String(percent).split('.')[1] || '').length;
-        const startTime = performance.now();
-        let countValue = 0;
-
-        const countUp = (timeStamp: any) => {
-          const elapsed = timeStamp - startTime;
-          countValue = (elapsed / duration) * percent;
-          eleProgressValue.innerText = countValue.toFixed(deciamlPointLength);
-
-          if (elapsed < duration) {
-            requestAnimationFrame(countUp);
-          } else {
-            eleProgressValue.innerText = percent;
-          }
-        };
-
-        // Styles for progress bar animation
-        entry.target.style.cssText = `
-      --duration: ${duration}ms;
-      --start-rotate: ${startPosition};
-      --stroke-dashoffset: ${strokeDashOffset};
-      --stroke-color: ${entry.target.dataset.strokeColor};
-      --stroke-width: ${entry.target.dataset.strokeWidth};
-    `;
-
-        requestAnimationFrame(countUp);
-
-        entry.target.classList.add('is-visible');
-      });
-    };
-
-    const options = {
-      root: root,
-      rootMargin: '78px 0px -40% 0px',
-      threshold: 0.0,
-    };
-
-    const obsever = new IntersectionObserver(observeAction, options);
-
-    progressItems.forEach((target) => {
-      obsever.observe(target);
-    });
-  }
+  useIntersectionObserver([s1, s2, s3, s4], showSlils, skilsOptions);
   return (
     <div
       id="skils"
@@ -125,9 +48,9 @@ const Skils: FC<TopProps> = ({ item }) => {
           </p>
         </div>
         {/* progressbar */}
-
         <div className="progress-container">
           <div
+            ref={s1}
             className="progress-item"
             data-percent="60"
             data-duration="2200"
@@ -158,6 +81,7 @@ const Skils: FC<TopProps> = ({ item }) => {
             <div className="progress-title">React</div>
           </div>
           <div
+            ref={s2}
             className="progress-item"
             data-percent="30"
             data-duration="3200"
@@ -188,6 +112,7 @@ const Skils: FC<TopProps> = ({ item }) => {
             <div className="progress-title">TypeScript</div>
           </div>
           <div
+            ref={s3}
             className="progress-item"
             data-percent="70"
             data-duration="4200"
@@ -218,6 +143,7 @@ const Skils: FC<TopProps> = ({ item }) => {
             <div className="progress-title">PHP</div>
           </div>
           <div
+            ref={s4}
             className="progress-item"
             data-percent="40"
             data-duration="5800"
